Hide Application template tab if not capable
