+++ public/controller-config.js
@@ -1,8 +1,10 @@
 window.controllerConfig = {
+  ip: '0.0.0.0',
   port: '51121',
   refresh: 3000,
+  dev: true,
   user: {
-    email: '',
-    password: ''
+    email: 'user@domain.com',
+    password: '#Bugs4Fun'
   }
 }
